Moved enet-multiplayer specific stuff from header to cpp file.
